+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Canonical reference for changes, improvements, and bugfixes for Boundary.
 
 ## Next
 
-* Remove deprecated `controllers` field from the worker config, which was deprecated for
+* Remove deprecated `controllers` field from the worker config, which was deprecated in 0.9.0 for
 `initial_upstreams`([PR](https://github.com/hashicorp/boundary/pull/5125))
 
 ## 0.17.2 (Unreleased)